Structural corrections
Delete spaces after <A> and <hi> opening tags and before </A> and </hi> closing tags.
Add missing spaces before <A> and <hi> opening tags and after </A> and </hi> closing tags (or the related brackets and interpunctuation).
Add missing spaces before † , ‡ and ↓.
Add additional <hi> and </hi> tags to ensure that brackets and arabic text are not italicized (to prevent nesting errors).
Replace { and } by the correct brackets - usually [ and ].

Ensure that numbers of verb forms are followed by "." if an arabic text is following and by ":" if "see" is following (reference). For consitency all "see" after ":" are written with small "s".

In some locations entries were split into two (i. e. 1. and 2.) This will be continued in the next commit.