Add "ref" elements for references to other roots
Added "<ref>" elements for nearly all references in the form:
     see art. <A>...</A>

This was done also if "See" is written with capital S and if line breaks are included within the reference).

References were only added of the target ID exists.
In several locations the text had to be "regularized" for this purpose
(marked by <choice><orig>...</orig><reg resp="#WHL">...</reg></choice>)
Only in seven locations references were not addad as the target roots could not be found/resolved.

All Arabic text within these references was transliterated to Arabic.