Add further "ref" elements for "in art." references
Added "<ref>" elements for most references in the form:
in art. <A>...</A>

This was done also if line breaks are included within the reference).

References were only added if the target ID exists.
Regularization of "broken" references was not yet done. For those refrerences no "ref" tags were added.

All Arabic text within these references was transliterated to Arabic.