Add further "<ref>" tags for "of art." references
Added "<ref>" elements for nearly all references in the form:
of art. <A>...</A>

This was done also if line breaks are included within the reference).

References were only added if the target ID exists.
Regularization of 5 "broken" references was not yet done. For those refrerences no "ref" tags were added.

All Arabic text within these references was transliterated to Arabic.