Added references to verb forms in other articles
Added missing <form> elements and "xml:id" for verb form entries with "[" at the beginning.

Added all crossreferences to verb forms in other articles which could be identified.
In a few locations corrections had to be made, which are marked as such.

"and" or "but" in front of a <ref> tag were moved behind the <ref> tag, as they are pert of the reference.